完成 index 的重构
