+++ package.json
@@ -4,12 +4,12 @@
   "productName": "Majsoul Plus",
   "author": "MajsoulPlus Team",
   "description": "Majsoul Plus",
-  "main": "./dist/index-refactor.js",
+  "main": "./dist/index.js",
   "scripts": {
     "sass": "node-sass assets/manager/ -o assets/manager/",
     "pre-build": "tsc && yarn sass && node build.js",
-    "debug": "cross-env NODE_ENV=development electron ./dist/index-refactor.js",
-    "debug:detailed": "cross-env NODE_ENV=development electron ./dist/index-refactor.js --enable-logging --v=1 --disable-gpu",
+    "debug": "cross-env NODE_ENV=development electron .",
+    "debug:detailed": "cross-env NODE_ENV=development electron . --enable-logging --v=1 --disable-gpu",
     "start": "yarn pre-build && yarn debug",
     "start:detailed": "yarn pre-build && yarn debug:detailed",
     "pack": "npm run lint && npm run build",
