+++ src/bin/main/mainLoader.ts
@@ -1,222 +1,29 @@
 import { ipcRenderer, screen as electronScreen } from 'electron'
-import * as fs from 'fs'
-import * as path from 'path'
 import i18n from '../../i18n'
-import { Global } from '../../global'
+import { Global, RemoteDomains } from '../../global'
+import { MajsoulPlus } from '../../majsoul_plus'
 
-const userConfigs = require(Global.UserConfigPath)
+const userConfigs: MajsoulPlus.UserConfig = require(Global.UserConfigPath)
 
-const mainWindow: Electron.WebviewTag = document.querySelector(
-  '#mainWindow'
-) as Electron.WebviewTag
-const mainWindowBox: HTMLDivElement = document.querySelector(
-  '#mainWindowBox'
-) as HTMLDivElement
-
+const mainWindow: Electron.WebviewTag = document.querySelector('#mainWindow')
+const mainWindowBox: HTMLDivElement = document.querySelector('#mainWindowBox')
 const scalePercent = userConfigs.window.renderingMultiple
 
 let webContents: Electron.webContents
-
-let executeScriptsCodes: string[] = []
-
-let serverPort: number
-
-let clientRect: DOMRect
-
-const prebuildExecuteCode = executeScriptInfo => {
-  const executePreferences = executeScriptInfo.executePreferences
-    ? executeScriptInfo.executePreferences
-    : {}
-  let codeEntry = executeScriptInfo.entry
-  if (!codeEntry) {
-    codeEntry = 'script.js'
-  }
-  let code = ''
-  if (Array.isArray(codeEntry)) {
-    codeEntry.forEach(entry => {
-      code +=
-        '\n' +
-        fs
-          .readFileSync(path.join(executeScriptInfo.filesDir, entry))
-          .toString('utf-8')
-    })
-  } else {
-    code = fs
-      .readFileSync(path.join(executeScriptInfo.filesDir, codeEntry))
-      .toString('utf-8')
-  }
-  const sanboxCode = `const sandbox = new Proxy({}, {
-    get(target, prop) {
-      const eventHandlers = []
-      if (target.hasOwnProperty(prop)) {
-        return target[prop]
-      }
-      if (prop === "window") {
-        return sandbox
-      }
-      if (prop === "global") {
-        return sandbox
-      }
-      if (prop === "open") {
-        return window.open.bind(window)
-      }
-      if (prop === "setTimeout") {
-        return (fun, ...props)=>{return setTimeout(fun.bind(sandbox), ...props)}
-      }
-      if (prop === "clearTimeout") {
-        return (...props)=>{return clearTimeout(...props).bind(window)}
-      }
-      if (prop === "setInterval") {
-        return (fun, ...props)=>{return setInterval(fun.bind(sandbox), ...props)}
-      }
-      if (prop === "clearInterval") {
-        return (...props)=>{return clearInterval(...props).bind(window)}
-      }
-      if (prop === "requestAnimationFrame") {
-        return (fun, ...props)=>{return requestAnimationFrame(fun.bind(sandbox), ...props)}
-      }
-      if (prop === "cancelAnimationFrame") {
-        return (...props)=>{return cancelAnimationFrame(...props).bind(window)}
-      }
-      if (prop === "addEventListener") {
-        return (evt, fun, ...props)=>{
-          const bindFun = fun.bind(sandbox)
-          eventHandlers.push({
-            unbind: fun,
-            bind: bindFun
-          })
-          window.addEventListener(evt, bindFun, ...props)
-        }
-      }
-      if (prop === "removeEventListener") {
-        return (evt, fun, ...props)=>{
-          let flag = false
-          eventHandlers.some((element,index,arr)=>{
-            const found = element.unbind === fun
-            if(found){
-              flag = window.removeEventListener(evt, bindFun, ...props)
-              arr.splice(index,1)
-              return true
-            }else{
-              return false
-            }
-          })
-        }
-      }
-      if (prop === "require" && ${!!executePreferences.nodeRequire} === false) {
-        return undefined
-      }
-      if (prop === "document" && ${!!executePreferences.document} === false) {
-        return undefined
-      }
-      if (prop === "localStorage" && ${!!executePreferences.localStorage} === false) {
-        return undefined
-      }
-      if (prop === "XMLHttpRequest" && ${!!executePreferences.XMLHttpRequest} === false) {
-        return undefined
-      }
-      if (prop === "WebSocket" && ${!!executePreferences.WebSocket} === false) {
-        return undefined
-      }
-      if (prop === Symbol.unscopables) {
-        return undefined
-      }
-      return window[prop]
-    },
-    has() { return true },
-    set(target, prop, value) {
-      target[prop] = value
-      if (${!!executePreferences.writeableWindowObject} === true) {
-        window[prop] = target[prop]
-      }
-      return target[prop]
-    }
-  })`
-  if (!executeScriptInfo.sync) {
-    code = `(()=>{
-      let __raf;
-      ${sanboxCode}
-      const __rafFun=()=>{
-        if(window.game) {
-              with (sandbox) {
-              ${code}
-            }
-          } else
-          {__raf=requestAnimationFrame(__rafFun)
-          }
-        }
-      __raf=requestAnimationFrame(__rafFun)
-      })()`
-  } else {
-    code = `(()=>{
-      ${sanboxCode}
-      with (sandbox) {${code}}
-    })()`
-  }
-  return code
-}
-
-// Sanbox from https://zhuanlan.zhihu.com/p/58602800
-const sandbox = new Proxy(
-  {},
-  {
-    get(target, prop) {
-      if (target.hasOwnProperty(prop)) {
-        return target[prop]
-      }
-      if (prop === 'window') {
-        return sandbox
-      }
-      if (prop === 'global') {
-        return sandbox
-      }
-      if (prop === 'require' && false === false) {
-        return undefined
-      }
-      if (prop === 'document' && false === false) {
-        return undefined
-      }
-      if (prop === 'localStorage' && false === false) {
-        return undefined
-      }
-      if (prop === 'XMLHttpRequest' && false === false) {
-        return undefined
-      }
-      if (prop === 'WebSocket' && false === false) {
-        return undefined
-      }
-      if (prop === Symbol.unscopables) {
-        return undefined
-      }
-      return window[prop]
-    },
-    has() {
-      return true
-    },
-    set(target, prop, value) {
-      target[prop] = value
-
-      // FIXME: Useless true === true
-      // if (true === true) {
-      window[prop] = target[prop]
-      // }
-      return true
-    }
-  }
-)
+let clientRect: Electron.Rectangle
 
 let screenshotCounter = 0
 let screenshotTimer: NodeJS.Timeout
-const showScreenshotLabel = src => {
+function showScreenshotLabel(src: string) {
   const screenshotImage: HTMLImageElement = document.querySelector(
     '#screenshotImage'
-  ) as HTMLImageElement
-  const screenshotText: HTMLParagraphElement = document.getElementById(
-    'screenshotText'
-  ) as HTMLParagraphElement
-  const screenshotLabel: HTMLDivElement = document.getElementById(
-    'screenshotLabel'
-  ) as HTMLDivElement
+  )
+  const screenshotText: HTMLParagraphElement = document.querySelector(
+    '#screenshotText'
+  )
+  const screenshotLabel: HTMLDivElement = document.querySelector(
+    '#screenshotLabel'
+  )
   screenshotImage.src = src
   screenshotText.innerText = screenshotCounter++
     ? i18n.text.main.screenshotsSaved(screenshotCounter)
@@ -234,8 +41,8 @@ const showScreenshotLabel = src => {
   }, 8000)
 }
 
-ipcRenderer.on('window-resize', (event, ...args) => {
-  clientRect = args[0]
+ipcRenderer.on('window-resize', (event, rect: Electron.Rectangle) => {
+  clientRect = rect
 })
 
 ipcRenderer.on('take-screenshot', () => {
@@ -263,58 +70,51 @@ ipcRenderer.on('take-screenshot', () => {
   }
 })
 
-ipcRenderer.on('screenshot-saved', (event, ...args) => {
-  const src = args[0]
-  showScreenshotLabel('file://' + src)
-})
+ipcRenderer.on(
+  'screenshot-saved',
+  (event: Electron.Event, filePath: string) => {
+    showScreenshotLabel('file://' + filePath)
+  }
+)
 
 ipcRenderer.on('open-devtools', () => {
   if (webContents) {
-    // webContents.openDevTools({ mode: "detach" })
     mainWindow.openDevTools()
   }
 })
 
-const testRedirectGameWindow = url => {
+let serverInfo: {
+  url: string
+  port: number
+  http: boolean
+}
+
+function isVanillaGameUrl(url: string) {
   return (
-    url.startsWith(Global.RemoteDomain) ||
-    url.startsWith(Global.HttpRemoteDomain)
+    url.startsWith(RemoteDomains[0].domain) ||
+    url.startsWith(RemoteDomains[1].domain) ||
+    url.startsWith(RemoteDomains[2].domain)
   )
 }
 
-const testIsLocalGameWindow = url => {
-  return url.startsWith('https://localhost:')
+function isLocalHost(url: string) {
+  return url.match(/https?:\/\/localhost:/)
 }
-const getLocalUrlWithParams = url => {
+
+function getLocalUrlWithParams(url: string) {
   if (url.includes('?')) {
-    return `https://localhost:${serverPort}/${url.substring(url.indexOf('?'))}`
+    return `${serverInfo.url}${url.substring(url.indexOf('?'))}`
   }
-  return `https://localhost:${serverPort}/`
+  return serverInfo.url
 }
-const redirectGameWindow = (url: string, gameWindow) => {
+
+function redirectGameWindow(url: string, gameWindow: Electron.WebviewTag) {
   const localUrl = getLocalUrlWithParams(url)
-  console.warn('Redirect Target:' + localUrl)
+  console.log('[Majsoul Plus] Redirect Target:' + localUrl)
   gameWindow.loadURL(localUrl)
 }
 
-ipcRenderer.on('server-port-load', (event, ...args) => {
-  console.warn('server-port-load')
-  serverPort = args[0]
-  ipcRenderer.send('server-port-loaded')
-})
-
-ipcRenderer.on('executes-load', (event, ...args) => {
-  console.warn('executes-load')
-  const executeScripts = args[0]
-  executeScriptsCodes = []
-  executeScripts.forEach(executeScript => {
-    const code = prebuildExecuteCode(executeScript)
-    executeScriptsCodes.push(code)
-  })
-  ipcRenderer.send('executes-loaded')
-})
-
-const scaleWindow = (percent = scalePercent) => {
+function scaleWindow(percent = scalePercent) {
   mainWindowBox.style.width = `${percent}vw`
   mainWindowBox.style.height = `${percent}vh`
   mainWindowBox.style.transform = `scale(${100 / percent}) translate(${(100 -
@@ -328,26 +128,19 @@ mainWindow.addEventListener('dom-ready', () => {
     webContents.setZoomFactor(1)
     ipcRenderer.send('main-loader-ready')
 
-    webContents.on('dom-ready', () => {
-      executeScriptsCodes.forEach(executeScriptCode => {
-        webContents.executeJavaScript(executeScriptCode)
-      })
-    })
-
-    webContents.on('will-navigate', (evt, url) => {
-      if (testRedirectGameWindow(url)) {
-        evt.preventDefault()
+    webContents.on('will-navigate', (event, url) => {
+      if (isVanillaGameUrl(url)) {
+        event.preventDefault()
         redirectGameWindow(url, mainWindow)
       }
     })
 
     if (process.env.NODE_ENV === 'development') {
-      // webContents.openDevTools({ mode: "detach" })
       mainWindow.openDevTools()
     }
   }
 
-  if (testIsLocalGameWindow(mainWindow.src)) {
+  if (isLocalHost(mainWindow.src)) {
     scaleWindow(scalePercent)
     mainWindow.insertCSS('body{overflow:hidden;}')
   } else {
@@ -355,15 +148,15 @@ mainWindow.addEventListener('dom-ready', () => {
   }
 })
 
-ipcRenderer.on('load-url', (event, ...args: string[]) => {
-  const url = args[0]
-  console.warn('LoadURL', url)
-  if (testRedirectGameWindow(url)) {
-    redirectGameWindow(url, mainWindow)
-  } else {
+ipcRenderer.on(
+  'load-url',
+  (event: Electron.Event, url: string, port: number, http: boolean) => {
+    serverInfo = { url, port, http }
+    console.log('[Majsoul Plus] LoadURL', serverInfo)
+
     mainWindow.loadURL(url)
+    mainWindowBox.style.width = '100vw'
+    mainWindowBox.style.height = '100vh'
+    mainWindowBox.style.transform = 'none'
   }
-  mainWindowBox.style.width = '100vw'
-  mainWindowBox.style.height = '100vh'
-  mainWindowBox.style.transform = 'none'
-})
+)
